bpf: use configurable large buffers for HTTP requests

+++ bpf/generictracer/protocol_mysql.h
@@ -133,7 +133,7 @@ static __always_inline int mysql_read_fixup_buffer(const connection_info_t *conn
         bpf_dbg_printk("mysql_read_fixup_buffer: bug: mysql_buffer_size is not a power of 2");
         return -1;
     }
-    const u8 buf_len_mask = mysql_buffer_size - 1;
+    const u32 buf_len_mask = mysql_buffer_size - 1;
 
     struct mysql_state_data *state_data = bpf_map_lookup_elem(&mysql_state, conn_info);
     if (state_data != NULL) {
