+++ bpf/generictracer/protocol_postgres.h
@@ -62,7 +62,7 @@ static __always_inline int postgres_send_large_buffer(tcp_req_t *req,
         bpf_dbg_printk("postgres_send_large_buffer: bug: postgres_buffer_size is not a power of 2");
         return -1;
     }
-    const u8 buf_len_mask = postgres_buffer_size - 1;
+    const u32 buf_len_mask = postgres_buffer_size - 1;
 
     tcp_large_buffer_t *large_buf = (tcp_large_buffer_t *)postgres_large_buffers_mem();
     if (!large_buf) {
